requirements.txt
requests>=2.31.0
pandas>=1.5.0
pyspark>=3.4.0
delta-spark>=2.4.0
streamlit>=1.28.0
plotly>=5.17.0
yfinance>=0.2.28
alpha-vantage>=2.3.1